Microservicios/Microservicios/instalaciones_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
Flask-Cors==3.0.10
marshmallow==3.21.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow-sqlalchemy

docxtpl==0.16.7
python-docx==1.1.0

requests==2.31.0
pytz==2024.1
python-dateutil==2.9.0.post0

twilio
sendgrid